requirements.txt
pydantic>=2.0.0
sentence-transformers>=2.2.2
transformers==4.33.0 
gunicorn

faiss-cpu==1.7.4
# numpy==1.26.4
requests==2.31.0
Flask==2.2.5
langchain==0.1.13
langchain-community>=0.0.29
google-generativeai==0.3.2
